requirements.txt
aiogram==2.21
mcrcon==0.7.0
python-dotenv==0.20.0
